+++ CMakeFiles/gbui_meta.dir/cmake_clean.cmake
@@ -1,0 +1,11 @@
+file(REMOVE_RECURSE
+  "CMakeFiles/gbui_meta.dir/src/meta/components.cpp.o"
+  "CMakeFiles/gbui_meta.dir/src/meta/components.cpp.o.d"
+  "libgbui_meta.a"
+  "libgbui_meta.pdb"
+)
+
+# Per-language clean rules from dependency scanning.
+foreach(lang CXX)
+  include(CMakeFiles/gbui_meta.dir/cmake_clean_${lang}.cmake OPTIONAL)
+endforeach()