+++ CTestTestfile.cmake
@@ -1,0 +1,9 @@
+# CMake generated Testfile for 
+# Source directory: /home/gustavo/Desktop/projects/gitbox/gbui
+# Build directory: /home/gustavo/Desktop/projects/gitbox/gbui
+# 
+# This file includes the relevant testing commands required for 
+# testing this directory and lists subdirectories to be tested as well.
+subdirs("tests")
+subdirs("examples")
+subdirs("demos")