+++ demos/CMakeFiles/gbui_demos.dir/DependInfo.cmake
@@ -1,0 +1,37 @@
+
+# Consider dependencies only in project.
+set(CMAKE_DEPENDS_IN_PROJECT_ONLY OFF)
+
+# The set of languages for which implicit dependencies are needed:
+set(CMAKE_DEPENDS_LANGUAGES
+  )
+
+# The set of dependency files which are needed:
+set(CMAKE_DEPENDS_DEPENDENCY_FILES
+  "/home/gustavo/Desktop/projects/gitbox/gbui/demos/src/analytics.cpp" "demos/CMakeFiles/gbui_demos.dir/src/analytics.cpp.o" "gcc" "demos/CMakeFiles/gbui_demos.dir/src/analytics.cpp.o.d"
+  "/home/gustavo/Desktop/projects/gitbox/gbui/demos/src/catalog/charts.cpp" "demos/CMakeFiles/gbui_demos.dir/src/catalog/charts.cpp.o" "gcc" "demos/CMakeFiles/gbui_demos.dir/src/catalog/charts.cpp.o.d"
+  "/home/gustavo/Desktop/projects/gitbox/gbui/demos/src/catalog/components.cpp" "demos/CMakeFiles/gbui_demos.dir/src/catalog/components.cpp.o" "gcc" "demos/CMakeFiles/gbui_demos.dir/src/catalog/components.cpp.o.d"
+  "/home/gustavo/Desktop/projects/gitbox/gbui/demos/src/catalog/containers.cpp" "demos/CMakeFiles/gbui_demos.dir/src/catalog/containers.cpp.o" "gcc" "demos/CMakeFiles/gbui_demos.dir/src/catalog/containers.cpp.o.d"
+  "/home/gustavo/Desktop/projects/gitbox/gbui/demos/src/catalog/controls.cpp" "demos/CMakeFiles/gbui_demos.dir/src/catalog/controls.cpp.o" "gcc" "demos/CMakeFiles/gbui_demos.dir/src/catalog/controls.cpp.o.d"
+  "/home/gustavo/Desktop/projects/gitbox/gbui/demos/src/catalog/overlays.cpp" "demos/CMakeFiles/gbui_demos.dir/src/catalog/overlays.cpp.o" "gcc" "demos/CMakeFiles/gbui_demos.dir/src/catalog/overlays.cpp.o.d"
+  "/home/gustavo/Desktop/projects/gitbox/gbui/demos/src/catalog/registry.cpp" "demos/CMakeFiles/gbui_demos.dir/src/catalog/registry.cpp.o" "gcc" "demos/CMakeFiles/gbui_demos.dir/src/catalog/registry.cpp.o.d"
+  "/home/gustavo/Desktop/projects/gitbox/gbui/demos/src/grid.cpp" "demos/CMakeFiles/gbui_demos.dir/src/grid.cpp.o" "gcc" "demos/CMakeFiles/gbui_demos.dir/src/grid.cpp.o.d"
+  "/home/gustavo/Desktop/projects/gitbox/gbui/demos/src/host.cpp" "demos/CMakeFiles/gbui_demos.dir/src/host.cpp.o" "gcc" "demos/CMakeFiles/gbui_demos.dir/src/host.cpp.o.d"
+  "/home/gustavo/Desktop/projects/gitbox/gbui/demos/src/logistics.cpp" "demos/CMakeFiles/gbui_demos.dir/src/logistics.cpp.o" "gcc" "demos/CMakeFiles/gbui_demos.dir/src/logistics.cpp.o.d"
+  "/home/gustavo/Desktop/projects/gitbox/gbui/demos/src/markets.cpp" "demos/CMakeFiles/gbui_demos.dir/src/markets.cpp.o" "gcc" "demos/CMakeFiles/gbui_demos.dir/src/markets.cpp.o.d"
+  "/home/gustavo/Desktop/projects/gitbox/gbui/demos/src/production.cpp" "demos/CMakeFiles/gbui_demos.dir/src/production.cpp.o" "gcc" "demos/CMakeFiles/gbui_demos.dir/src/production.cpp.o.d"
+  "/home/gustavo/Desktop/projects/gitbox/gbui/demos/src/registry.cpp" "demos/CMakeFiles/gbui_demos.dir/src/registry.cpp.o" "gcc" "demos/CMakeFiles/gbui_demos.dir/src/registry.cpp.o.d"
+  "/home/gustavo/Desktop/projects/gitbox/gbui/demos/src/scada.cpp" "demos/CMakeFiles/gbui_demos.dir/src/scada.cpp.o" "gcc" "demos/CMakeFiles/gbui_demos.dir/src/scada.cpp.o.d"
+  "/home/gustavo/Desktop/projects/gitbox/gbui/demos/src/weather.cpp" "demos/CMakeFiles/gbui_demos.dir/src/weather.cpp.o" "gcc" "demos/CMakeFiles/gbui_demos.dir/src/weather.cpp.o.d"
+  )
+
+# Targets to which this target links which contain Fortran sources.
+set(CMAKE_Fortran_TARGET_LINKED_INFO_FILES
+  )
+
+# Targets to which this target links which contain Fortran sources.
+set(CMAKE_Fortran_TARGET_FORWARD_LINKED_INFO_FILES
+  )
+
+# Fortran module output directory.
+set(CMAKE_Fortran_TARGET_MODULE_DIR "")