+++ demos/CTestTestfile.cmake
@@ -1,0 +1,6 @@
+# CMake generated Testfile for 
+# Source directory: /home/gustavo/Desktop/projects/gitbox/gbui/demos
+# Build directory: /home/gustavo/Desktop/projects/gitbox/gbui/demos
+# 
+# This file includes the relevant testing commands required for 
+# testing this directory and lists subdirectories to be tested as well.