+++ examples/CMakeFiles/gbui_controls.dir/cmake_clean.cmake
@@ -1,0 +1,12 @@
+file(REMOVE_RECURSE
+  "CMakeFiles/gbui_controls.dir/link.d"
+  "CMakeFiles/gbui_controls.dir/controls/main.cpp.o"
+  "CMakeFiles/gbui_controls.dir/controls/main.cpp.o.d"
+  "gbui_controls"
+  "gbui_controls.pdb"
+)
+
+# Per-language clean rules from dependency scanning.
+foreach(lang CXX)
+  include(CMakeFiles/gbui_controls.dir/cmake_clean_${lang}.cmake OPTIONAL)
+endforeach()