+++ examples/CMakeFiles/gbui_gallery.dir/cmake_clean.cmake
@@ -1,0 +1,12 @@
+file(REMOVE_RECURSE
+  "CMakeFiles/gbui_gallery.dir/link.d"
+  "CMakeFiles/gbui_gallery.dir/gallery/main.cpp.o"
+  "CMakeFiles/gbui_gallery.dir/gallery/main.cpp.o.d"
+  "gbui_gallery"
+  "gbui_gallery.pdb"
+)
+
+# Per-language clean rules from dependency scanning.
+foreach(lang CXX)
+  include(CMakeFiles/gbui_gallery.dir/cmake_clean_${lang}.cmake OPTIONAL)
+endforeach()